requirements.txt
azure-functions
requests
praw
pytz